refactor: Extract playerIdStorage utils from game socket store
- Add TypeDoc documentation for utility functions
- Improve code organization and maintainability
- Update imports and references
