+++ client/src/hooks/socket/useDrawingSocket.ts
@@ -4,9 +4,9 @@ import { useCallback, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import type { DrawUpdateResponse } from '@troublepainter/core';
 import type { DrawingData } from '@troublepainter/core';
-import { playerIdStorageUtils } from '@/stores/socket/gameSocket.store';
 import { SocketNamespace } from '@/stores/socket/socket.config';
 import { useSocketStore } from '@/stores/socket/socket.store';
+import { playerIdStorageUtils } from '@/utils/playerIdStorage';
 
 interface UseDrawingSocketProps {
   onDrawUpdate?: (response: DrawUpdateResponse) => void;
